[New] Adde test cases for transitions components date transitions summary
