+++ app/javascript/components/transitions/components/date-transitions-summary/component.jsx
@@ -31,7 +31,11 @@ const Component = ({ value }) => {
     );
   }
 
-  return <div className={css.date}>{i18n.localizeDate(value)}</div>;
+  return (
+    <div className={css.date} data-testid="localize-date">
+      {i18n.localizeDate(value)}
+    </div>
+  );
 };
 
 Component.displayName = NAME;
